sm8150-common: Re-adjust ACDB IDs like stock
After further investigation, it seems that the ACDB ID for
SND_DEVICE_IN_HANDSET_MIC results wrong when compared to a stock-based GSI
image and edo.

As for video recording, it turns out that instead of the camcorder-mic
path, LOS likes using handset-stereo-dmic-ef whose ACDB (according to edo)
should be 35: adjust the configs for these paths, for optimal audio volume.

Fixes: 385bc9fddde6c9c58d1e ("sm8150-common: Fix recording ABCD ID")
Fixes: 1df4d13faf7c099dfd19 ("sm8150-common: Bump some microphone volumes")
Signed-off-by: voidanix <voidanix@keyedlimepie.org>